requirements.txt
# Core ML
datasets==4.8.0
transformers>=4.40.0
peft==0.11.1
bitsandbytes==0.43.1
accelerate==0.30.1

# Evaluation
rouge-score==0.1.2

# Experiment tracking
mlflow==2.13.0

# Demo UI
gradio==4.31.5

# Utilities
python-dotenv==1.0.1
